src/main/java/shop/mtcoding/blog/board/BoardPersistRepository.java
package shop.mtcoding.blog.board;

import jakarta.persistence.EntityManager;
import jakarta.persistence.Query;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Repository;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@RequiredArgsConstructor
@Repository
public class BoardPersistRepository {
    private final EntityManager em;

    public Board findById(Integer id) {
        Board board = em.find(Board.class, id);

        return board;
    }

    public List<Board> findAll() {
        Query query = em.createQuery("select b from Board b order by b.id desc", Board.class);

        return query.getResultList();
    }

    @Transactional
    public Board save(Board board) {

        // 1. 비영속 객체 생성
        em.persist(board);

        // board -> 영속 객체
        return board;
    }
}
