src/test/java/shop/mtcoding/blog/board/BoardPersistRepositoryTest.java
package shop.mtcoding.blog.board;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
import org.springframework.context.annotation.Import;

import javax.swing.*;
import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;

@Import(BoardPersistRepository.class)
@DataJpaTest
public class BoardPersistRepositoryTest {

    @Autowired
    private BoardPersistRepository boardPersistRepository;

    @Test
    public void save_test() {
        // given
        Board board = new Board("제목5", "내용5", "ssar");

        boardPersistRepository.save(board);
        System.out.println("save_test : " + board);

        // then
    }

    @Test
    public void findAll_test() {
        // given

        // when
        List<Board> boardList = boardPersistRepository.findAll();

        // then
        System.out.println("findAll_size : " + boardList.size());

        assertThat(boardList.size()).isEqualTo(4);
    }

    @Test
    public void findById_test() {
        Integer id = 1;

        Board board = boardPersistRepository.findById(id);
        boardPersistRepository.findById(id);
    }
}
